+++ bin/miden-cli/src/errors.rs
@@ -49,9 +49,10 @@ pub enum CliError {
     #[diagnostic(
         code(cli::config_error),
         help(
-            "Check if the configuration file exists and is well-formed. If it does not exist, run `{} init` command to create it.",
+            "Check if the configuration file exists and is well-formed. If it does not exist, run `{} init` command to create it. \
+            If it already exists, run `{} clear-config` to remove it.",
+            client_binary_name().display(),
             client_binary_name().display()
-
         )
     )]
     Config(#[source] SourceError, String),